+++ ActionHandler/PubliccodeFindRepositoriesThroughOrganizationHandler.php
@@ -2,16 +2,16 @@
 
 namespace OpenCatalogi\OpenCatalogiBundle\ActionHandler;
 
-use  OpenCatalogi\OpenCatalogiBundle\Service\PubliccodeService;
+use OpenCatalogi\OpenCatalogiBundle\Service\FindRepositoriesThroughOrganizationService;
 use CommonGateway\CoreBundle\ActionHandler\ActionHandlerInterface;
 
 class PubliccodeFindRepositoriesThroughOrganizationHandler implements ActionHandlerInterface
 {
-    private PubliccodeService $publiccodeService;
+    private FindRepositoriesThroughOrganizationService $findRepositoriesThroughOrganizationService;
 
-    public function __construct(PubliccodeService $publiccodeService)
+    public function __construct(FindRepositoriesThroughOrganizationService $findRepositoriesThroughOrganizationService)
     {
-        $this->publiccodeService = $publiccodeService;
+        $this->findRepositoriesThroughOrganizationService = $findRepositoriesThroughOrganizationService;
     }
 
     public function getConfiguration()
@@ -36,6 +36,6 @@ class PubliccodeFindRepositoriesThroughOrganizationHandler implements ActionHand
 
     public function run(array $data, array $configuration): array
     {
-        return $this->publiccodeService->enrichOrganizationWithRepositoriesHandler($data, $configuration);
+        return $this->findRepositoriesThroughOrganizationService->enrichOrganizationWithRepositoriesHandler($data, $configuration);
     }
 }
