+++ Service/CheckRepositoriesForPubliccodeService.php
@@ -1,0 +1,119 @@
+<?php
+
+namespace OpenCatalogi\OpenCatalogiBundle\Service;
+
+use App\Entity\Entity;
+use App\Entity\ObjectEntity;
+use Doctrine\ORM\EntityManagerInterface;
+use Exception;
+use GuzzleHttp\Exception\GuzzleException;
+use Symfony\Component\HttpFoundation\Response;
+
+class CheckRepositoriesForPubliccodeService
+{
+    private EntityManagerInterface $entityManager;
+    private GithubApiService $githubService;
+    private array $configuration;
+    private array $data;
+
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        GithubApiService $githubService
+    ) {
+        $this->entityManager = $entityManager;
+        $this->githubService = $githubService;
+        $this->configuration = [];
+        $this->data = [];
+    }
+
+    /**
+     * @param ObjectEntity $repository
+     * @param Entity       $componentEntity
+     * @param Entity       $descriptionEntity
+     * @param array        $publiccode
+     *
+     * @throws Exception
+     *
+     * @return ObjectEntity|null dataset at the end of the handler
+     */
+    public function enrichRepositoryWithPubliccode(ObjectEntity $repository, Entity $componentEntity, Entity $descriptionEntity, array $publiccode = []): ?ObjectEntity
+    {
+        if (!$repository->getValue('component')) {
+            $component = new ObjectEntity();
+            $component->setEntity($componentEntity);
+        } else {
+            $component = $repository->getValue('component');
+        }
+
+        if ($publiccode !== null) {
+            $component->setValue('softwareVersion', $publiccode['publiccodeYmlVersion'] ?? null);
+            $component->setValue('name', $publiccode['name'] ?? null);
+            $component->setValue('softwareType', $publiccode['softwareType'] ?? null);
+            $component->setValue('inputTypes', $publiccode['inputTypes'] ?? null);
+            $component->setValue('outputTypes', $publiccode['outputTypes'] ?? null);
+            $component->setValue('platforms', $publiccode['platforms'] ?? null);
+            $component->setValue('categories', $publiccode['categories'] ?? null);
+            $component->setValue('developmentStatus', $publiccode['developmentStatus'] ?? null);
+            $repository->setValue('url', $publiccode['url'] ?? null);
+
+            //            $component->setValue('releaseDate', $publiccode['releaseDate']);
+            //            $component->setValue('dependsOn', $publiccode['dependsOn']['open']['name']);
+            //            $component->setValue('dependsOn', $publiccode['dependsOn']['open']['versionMin']);
+
+            if (!$component->getValue('description')) {
+                $description = new ObjectEntity();
+                $description->setEntity($descriptionEntity);
+            } else {
+                $description = $component->getValue('description');
+            }
+
+            $description->setValue('shortDescription', $publiccode['description']['nl']['shortDescription'] ?? null);
+            $description->setValue('documentation', $publiccode['description']['nl']['documentation'] ?? null);
+            $description->setValue('apiDocumentation', $publiccode['description']['nl']['apiDocumentation'] ?? null);
+            $description->setValue('shortDescription', $publiccode['description']['en']['shortDescription'] ?? null);
+            $description->setValue('documentation', $publiccode['description']['en']['documentation'] ?? null);
+            $description->setValue('apiDocumentation', $publiccode['description']['en']['apiDocumentation'] ?? null);
+
+            $component->setValue('description', $description);
+            $this->entityManager->persist($description);
+
+            $component->setValue('url', $repository);
+            $this->entityManager->persist($component);
+            $repository->setValue('component', $component);
+            $this->entityManager->persist($repository);
+            $this->entityManager->flush();
+        } else {
+            $component->setValue('name', $repository->getValue('name'));
+        }
+
+        return $repository;
+    }
+
+    /**
+     * @param array $data          data set at the start of the handler
+     * @param array $configuration configuration of the action
+     *
+     * @throws GuzzleException
+     *
+     * @return array dataset at the end of the handler
+     */
+    public function enrichPubliccodeHandler(array $data, array $configuration): array
+    {
+        $this->configuration = $configuration;
+        $this->data = $data;
+
+        $repositoryEntity = $this->entityManager->getRepository('App:Entity')->find($this->configuration['repositoryEntityId']);
+        $componentEntity = $this->entityManager->getRepository('App:Entity')->find($this->configuration['componentEntityId']);
+        $descriptionEntity = $this->entityManager->getRepository('App:Entity')->find($this->configuration['descriptionEntityId']);
+        // If we want to do it for al repositories
+        foreach ($repositoryEntity->getObjectEntities() as $repository) {
+            if ($publiccodeUrl = $repository->getValue('publiccode_url')) {
+                if (is_array($publiccode = $this->githubService->getPubliccode($publiccodeUrl))) {
+                    $this->enrichRepositoryWithPubliccode($repository, $componentEntity, $descriptionEntity, $publiccode);
+                }
+            }
+        }
+
+        return $this->data;
+    }
+}
